rename data_structure_dicts.py to dicts.py
